Handle Windows command not found error being different
Using cross-spawn on Windows will return returnCode=1 and use stderr
message when command is not found. However it is expected that an Error is
returned from spawn in that case and error.code is ENOENT. This commit
handles Windows explicitly by checking if it is Windows, for the
returnCode, and also if the error message is appropriate.

Example Windows command not found error message:
	'command' is not recognized as an internal or external command,
	operable program or batch file.

/cc #397