Add ability to read rustfmt.toml
atom-beautifier runs rustfmt in a tmp folder and rustfmt with not pick up
the rustfmt.toml file in any of the crate root folders or its parent folders.
This commit adds the --config-path parameter with the current folder and will
load any rust format in the file's path

Also set the --write-output format to override from the default setting replace.
This will avoid creating an unnecessary backup file tmp source file