Fixes #924. Beautify On Save changes the TextEditor not the file
It seems that in a recent update of Atom, Atom-Beautify's beautify on
save feature would actually crash Atom and in some cases even delete all
of the contents of the file!

After some investigation, it looks like writing directly to the file
with fs.write may have been the culprit and crashing Atom.