Fixes #924. Prevent infinite loop with beautify on save
With the option `Ensure Single Trailing Newline` option enabled
in the Whitespace package, Atom would append a newline character after
beautification had occured and when the beautified code was saved it
would be changed and never settle on a stable text that would not change
the previously saved text in the editor. Thus an infinite loop on save.