Add support for Whalebrew installed executables
Whalebrew mounts a Docker volume to the host's current working directory (cwd)
therefore the cwd needs to be the temp directory and all file paths which
are to be beautified need to be given to Whalebrew as relative to cwd.
Often this means the temporary directory is simply removed from the prefix
of the file paths, and the cwd = the temporary directory.
See https://github.com/bfirsh/whalebrew#how-it-works for details.