Fixes #1728. Remove Shell-Env from Executable, use Atom's process.env instead
Atom has builtin a fix for the process.env being out of sync with
the user's actual environment variables. It is similar to how Shell-Env works
which Atom-Beautify was using. Now we no longer need it.

However, there is still an issue when running specs. Thus, Atom-Beautify
specs still patch process.env with Shell-Env.
See https://discuss.atom.io/t/specs-do-not-load-shell-environment-variables-activationhooks-core-loaded-shell-environment/44199
for details.