Fix blade whitespace issues
There are possible scenarions where
the workaround tag `<blade ** />` gets an
additional whitespace after going through
beautifyHTML(). This cases the tag to change
to `<blade ** / >`. This prevents
to revert the tag to `**`. This commit resolves
this issue, by extending the regex, so that
the blade tag will be caught, even if a whitespace
occurs.

Fixes #2414